refactor: pages 구조 변경
