+++ src/pages/surveys/create/card/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import CreateCardSurvey from "../../../components/surveys/create/card/CreateCardSurvey";
+import CreateCardSurvey from "../../../../components/surveys/create/card/CreateCardSurvey";
 
 export default function index() {
   return <CreateCardSurvey></CreateCardSurvey>;
