+++ src/pages/surveys/create/general/index.js
@@ -1,0 +1,6 @@
+import { React } from "react";
+import CreateSurvey from "../../../../components/surveys/create/general/CreateSurvey";
+
+export default function CreateSurvey2() {
+  return <CreateSurvey></CreateSurvey>;
+}
