+++ src/pages/surveys/inquire/created/card/index.js
