+++ src/pages/surveys/inquire/created/general/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import ShowSurvey from "../../../components/surveys/inquire/created/ShowSurvey";
+import ShowSurvey from "../../../../../components/surveys/inquire/created/ShowSurvey";
 
 export default function Show() {
   const { sharingKey } = useParams();
