+++ src/pages/surveys/inquire/submitted/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useParams } from "react-router-dom";
-import SubmittedSurvey from "../../../components/surveys/inquire/submitted/submittedSurvey";
+import SubmittedSurvey from "../../../../components/surveys/inquire/submitted/submittedSurvey";
 
 export default function InquireSubmittedSurvey() {
   const { sharingKey } = useParams();
