+++ src/pages/surveys/participate/index.js
@@ -1,8 +1,6 @@
-import { React, useEffect } from "react";
+import { React } from "react";
 import { useParams } from "react-router-dom";
 import SurveyDetail from "../../../components/surveys/participate/SurveyDetail";
-import { useRecoilValue } from "recoil";
-import { userState } from "../../../authentication/userState";
 
 export default function Participate() {
   const { sharingKey } = useParams();
